Fix lint

+++ crates/optimism/src/evm.rs
@@ -549,7 +549,7 @@ mod tests {
     #[test]
     #[cfg(feature = "blst")]
     fn test_halted_tx_call_bls12_381_pairing_out_of_gas() {
-        let pairing_gas: u64 = bls12_381_const::PAIRING_PAIRING_MULTIPLIER_BASE as u64
+        let pairing_gas: u64 = bls12_381_const::PAIRING_PAIRING_MULTIPLIER_BASE
             + bls12_381_const::PAIRING_PAIRING_OFFSET_BASE;
 
         let ctx = Context::op()
@@ -583,7 +583,7 @@ mod tests {
     #[test]
     #[cfg(feature = "blst")]
     fn test_tx_call_bls12_381_pairing_wrong_input_layout() {
-        let pairing_gas: u64 = bls12_381_const::PAIRING_PAIRING_MULTIPLIER_BASE as u64
+        let pairing_gas: u64 = bls12_381_const::PAIRING_PAIRING_MULTIPLIER_BASE
             + bls12_381_const::PAIRING_PAIRING_OFFSET_BASE;
 
         let ctx = Context::op()
